pyproject.toml
[tool.poetry]
name = "rinha-2-django"
version = "0.1.0"
description = ""
authors = ["Rafael Miranda <rafaelpadovezi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.2"
django-rest-framework = "^0.1.0"
psycopg = {extras = ["binary", "pool"], version = "^3.1.18"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
